[Protocol3] Added revert support + refactored block processing
